+++ _source/_assets/components/LittleLayoutField.spec.ts
@@ -1,0 +1,8 @@
+import { expect, it } from 'vitest'
+import { LittleLayoutField } from "./LittleLayoutField";
+
+it('test method', () => {
+    // const wrapper = new LittleLayoutField();
+    // expect(wrapper.testMethod()).toBe('hello');
+    // expect(wrapper.testMethod('hey')).toBe('hey');
+})