+++ _source/_assets/vite.config.ts
@@ -1,10 +1,16 @@
 import { defineConfig } from 'vite';
 import * as dotenv from 'dotenv';
+// import tailwindcss from 'tailwindcss';
 import vue from '@vitejs/plugin-vue';
 dotenv.config({ path: '../../assembly-line.env' });
 
 export default defineConfig({
   mode: 'development',
+  // css: {
+  //   postcss: {
+  //     plugins: [tailwindcss()],
+  //   }
+  // },
   build: {
     emptyOutDir: true,
     manifest: true,