+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.3.0.beta.0 - 2022-11-05
+### Changed
+- Improved performance by refactoring Little Layout field from a Vue app to a Web Component.
+- Updated JS bundler to Vite 4.x.
+- Updated Tailwind to Tailwind 3.x.
+- Bumped NodeJS to Node 18.x.
+
+### Removed
+- Removed message that popped up when it was detected that you would need to horizontally scroll.
+
+### Fixed
+- Removed field value change on initialization that caused Craft to think the field was updated before changes were made.
+
 
 ## 1.2.0 - 2022-05-01
 ### Added